Fix incorrect monaco reference in es-editor - add warnings if initialized incorrectly - inline code in initialize to prevent modules being placed throughout
