+++ .changeset/silly-goats-bathe.md
@@ -1,0 +1,5 @@
+---
+'@eventstore-ui/editor': patch
+---
+
+Bug Fix: Fix editor failing to initialize
