+++ packages/editor/.gitignore
@@ -2,6 +2,4 @@
 /initialize.js
 /configure.d.ts
 /configure.js
-/themes.d.ts
-/themes.js
 /workers
