+++ packages/editor/src/utils/globals.ts
@@ -1,0 +1,10 @@
+import type * as monaco from 'monaco-editor';
+
+export const MONACO = Symbol.for('monaco');
+
+declare global {
+    interface Window {
+        MonacoEnvironment: monaco.Environment;
+        [MONACO]: typeof monaco;
+    }
+}
