+++ packages/editor/src/utils/logger.ts
@@ -2,5 +2,5 @@ import { createLogger } from '@eventstore-ui/utils';
 
 export const logger = createLogger(
     '@eventstore-ui/editor',
-    'linear-gradient(to right, #2b5876 0%, #4e4376 51%, #2b5876 100%)',
+    'linear-gradient(90deg, #A100FFFF 0%, #71C4FFFF 100%)',
 );
